refactor(graph): 拆分 sigma-global-renderer 为 3 个 helper (#64 Phase1)
把云层几何/凸包/签名、SVG 渲染/overlay 工厂、世界↔屏幕坐标变换从
sigma-global-renderer.ts 抽到独立 helper：
- render/sigma-coordinates.ts
- render/community-cloud-geometry.ts
- render/sigma-overlay-svg.ts

主文件 1386 → 1065 行。纯搬迁、零行为变化：运行时依赖单向（主文件→helper），
仅类型用 import type 回引；公开 API 与 barrel 不变。typecheck + 417 测试全绿。
