+++ packages/graph-engine/src/render/sigma-global-renderer.ts
@@ -19,10 +19,27 @@ import {
   type SigmaGlobalRenderedObject,
   type SigmaGlobalNodeDragSession
 } from "./sigma-global-drag";
-import { rootClientPointToScreenPoint, screenPointToWorldPoint, worldPointToCssPercentPoint, type GraphScreenPoint } from "./geometry";
+import { screenPointToWorldPoint, type GraphScreenPoint } from "./geometry";
 import { graphSpatialHitToGestureTarget, type GraphGestureTarget } from "./gestures";
 import { edgeRelationClass } from "./model";
 import { DEFAULT_RENDERER_VIEWPORT, type RendererViewport, type RendererViewportSize } from "./viewport";
+import { overlayPointerScreenPoint, sigmaScreenPointToWorldPoint, sigmaWorldPointToScreenPoint } from "./sigma-coordinates";
+import {
+  sigmaCommunityCloud,
+  sigmaCommunityCloudBasisById,
+  sigmaCommunityCloudBasisByIdWithNodePoint,
+  sigmaCommunityCloudBasisByIdWithReuse,
+  sigmaProjectedCloudHullPoints
+} from "./community-cloud-geometry";
+import {
+  applyOverlayBox,
+  createSigmaOverlayRoot,
+  nextSigmaCloudFilterSequence,
+  sigmaCloudSvg,
+  sigmaOverlayButton,
+  sigmaOverlayPassiveElement,
+  sigmaSharedCloudFilterDef
+} from "./sigma-overlay-svg";
 
 export const SIGMA_GLOBAL_RENDERER_ID = "sigma-global" as const;
 
@@ -783,327 +800,6 @@ function createSigmaRoot(container: HTMLElement, theme: ThemeId): HTMLElement {
   root.tabIndex = 0;
   container.append(root);
   return root;
-}
-
-function createSigmaOverlayRoot(root: HTMLElement): HTMLElement {
-  const overlay = root.ownerDocument.createElement("div");
-  overlay.className = "sigma-global-overlay";
-  overlay.dataset.role = "sigma-global-overlay";
-  root.append(overlay);
-  return overlay;
-}
-
-function sigmaOverlayButton(ownerDocument: Document, kind: string, id: string, label: string): HTMLButtonElement {
-  const element = ownerDocument.createElement("button");
-  element.type = "button";
-  element.dataset.kind = kind;
-  element.dataset.id = id;
-  element.setAttribute("aria-label", label);
-  return element;
-}
-
-function sigmaOverlayPassiveElement(ownerDocument: Document, kind: string, id: string): HTMLDivElement {
-  const element = ownerDocument.createElement("div");
-  element.dataset.kind = kind;
-  element.dataset.id = id;
-  element.setAttribute("aria-hidden", "true");
-  element.tabIndex = -1;
-  element.style.pointerEvents = "none";
-  return element;
-}
-
-function applyOverlayBox(element: HTMLElement, box: { left: number; top: number; width: number; height: number }): void {
-  element.style.left = `${box.left}px`;
-  element.style.top = `${box.top}px`;
-  element.style.width = `${box.width}px`;
-  element.style.height = `${box.height}px`;
-}
-
-const SIGMA_OVERLAY_SVG_NS = "http://www.w3.org/2000/svg";
-
-interface SigmaCommunityCloud {
-  box: { left: number; top: number; width: number; height: number };
-  localPoints: Array<{ x: number; y: number }> | null;
-}
-
-interface SigmaCommunityCloudBasis {
-  hullPoints: Array<{ x: number; y: number }>;
-  signature: string;
-}
-
-function sigmaCommunityCloudBasisById(adapterData: GraphRendererAdapterData): Map<string, SigmaCommunityCloudBasis> {
-  const washByCommunityId = new Map(adapterData.renderable.communities.map((community) => [community.id, community.wash]));
-  const pointsByCommunityId = new Map<string, Array<{ x: number; y: number }>>();
-  for (const node of adapterData.nodes) {
-    if (!node.communityId) continue;
-    const wash = washByCommunityId.get(node.communityId);
-    if (!wash) continue;
-    const list = pointsByCommunityId.get(node.communityId);
-    const point = clampPointToWorldEllipse(node.point, wash);
-    if (list) list.push(point);
-    else pointsByCommunityId.set(node.communityId, [point]);
-  }
-  const output = new Map<string, SigmaCommunityCloudBasis>();
-  for (const [communityId, points] of pointsByCommunityId) {
-    output.set(communityId, { hullPoints: convexHull2d(points), signature: sigmaCommunityCloudSignature(points, washByCommunityId.get(communityId)) });
-  }
-  return output;
-}
-
-function sigmaCommunityCloudBasisByIdWithReuse(
-  previous: Map<string, SigmaCommunityCloudBasis>,
-  adapterData: GraphRendererAdapterData
-): Map<string, SigmaCommunityCloudBasis> {
-  const washByCommunityId = new Map(adapterData.renderable.communities.map((community) => [community.id, community.wash]));
-  const pointsByCommunityId = new Map<string, Array<{ x: number; y: number }>>();
-  for (const node of adapterData.nodes) {
-    if (!node.communityId) continue;
-    const wash = washByCommunityId.get(node.communityId);
-    if (!wash) continue;
-    const list = pointsByCommunityId.get(node.communityId);
-    const point = clampPointToWorldEllipse(node.point, wash);
-    if (list) list.push(point);
-    else pointsByCommunityId.set(node.communityId, [point]);
-  }
-  const output = new Map<string, SigmaCommunityCloudBasis>();
-  for (const [communityId, points] of pointsByCommunityId) {
-    const signature = sigmaCommunityCloudSignature(points, washByCommunityId.get(communityId));
-    const cached = previous.get(communityId);
-    output.set(communityId, cached?.signature === signature ? cached : { hullPoints: convexHull2d(points), signature });
-  }
-  return output;
-}
-
-function sigmaCommunityCloudBasisByIdWithNodePoint(
-  previous: Map<string, SigmaCommunityCloudBasis>,
-  adapterData: GraphRendererAdapterData,
-  nodeId: string
-): Map<string, SigmaCommunityCloudBasis> {
-  const changedNode = adapterData.nodes.find((node) => node.id === nodeId);
-  if (!changedNode?.communityId) return previous;
-  const community = adapterData.renderable.communities.find((item) => item.id === changedNode.communityId);
-  if (!community?.wash) return previous;
-  const wash = community.wash;
-  const points = adapterData.nodes
-    .filter((node) => node.communityId === changedNode.communityId)
-    .map((node) => clampPointToWorldEllipse(node.point, wash));
-  const signature = sigmaCommunityCloudSignature(points, wash);
-  const cached = previous.get(changedNode.communityId);
-  if (cached?.signature === signature) return previous;
-  const next = new Map(previous);
-  next.set(changedNode.communityId, { hullPoints: convexHull2d(points), signature });
-  return next;
-}
-
-function sigmaCommunityCloudSignature(
-  points: readonly { x: number; y: number }[],
-  wash: { cx: number; cy: number; rx: number; ry: number } | null | undefined
-): string {
-  const parts = wash ? [wash.cx, wash.cy, wash.rx, wash.ry] : [];
-  for (const point of points) parts.push(point.x, point.y);
-  return parts.map((value) => String(Math.round(value * 1000) / 1000)).join(",");
-}
-
-function sigmaProjectedCloudHullPoints(
-  basis: SigmaCommunityCloudBasis | undefined,
-  sigma: SigmaGlobalSigmaLike,
-  options: Pick<SigmaGlobalRendererCreateOptions, "viewport" | "viewportSize" | "adapterData">
-): GraphScreenPoint[] {
-  return basis?.hullPoints.map((point) => sigmaWorldPointToScreenPoint(sigma, point, options)) ?? [];
-}
-
-function clampPointToWorldEllipse(
-  point: { x: number; y: number },
-  ellipse: { cx: number; cy: number; rx: number; ry: number }
-): { x: number; y: number } {
-  const rx = Math.max(1, ellipse.rx);
-  const ry = Math.max(1, ellipse.ry);
-  const dx = point.x - ellipse.cx;
-  const dy = point.y - ellipse.cy;
-  const distance = Math.hypot(dx / rx, dy / ry);
-  if (distance <= 1) return { x: point.x, y: point.y };
-  return {
-    x: ellipse.cx + dx / distance,
-    y: ellipse.cy + dy / distance
-  };
-}
-
-function sigmaCommunityCloud(
-  screenHullPoints: GraphScreenPoint[],
-  fallbackBox: { left: number; top: number; width: number; height: number }
-): SigmaCommunityCloud {
-  const hull = screenHullPoints;
-  if (hull.length >= 3) {
-    const cx = hull.reduce((sum, p) => sum + p.x, 0) / hull.length;
-    const cy = hull.reduce((sum, p) => sum + p.y, 0) / hull.length;
-    const expanded = hull.map((p) => clampPointToScreenEllipse({
-      x: p.x + (p.x - cx) * 0.4,
-      y: p.y + (p.y - cy) * 0.4
-    }, fallbackBox));
-    let minX = Infinity;
-    let minY = Infinity;
-    let maxX = -Infinity;
-    let maxY = -Infinity;
-    for (const p of expanded) {
-      minX = Math.min(minX, p.x);
-      minY = Math.min(minY, p.y);
-      maxX = Math.max(maxX, p.x);
-      maxY = Math.max(maxY, p.y);
-    }
-    const box = { left: minX, top: minY, width: Math.max(8, maxX - minX), height: Math.max(8, maxY - minY) };
-    return { box, localPoints: expanded.map((p) => ({ x: p.x - box.left, y: p.y - box.top })) };
-  }
-  return { box: fallbackBox, localPoints: null };
-}
-
-function clampPointToScreenEllipse(
-  point: GraphScreenPoint,
-  box: { left: number; top: number; width: number; height: number }
-): GraphScreenPoint {
-  const cx = box.left + box.width / 2;
-  const cy = box.top + box.height / 2;
-  const rx = Math.max(1, box.width / 2);
-  const ry = Math.max(1, box.height / 2);
-  const dx = point.x - cx;
-  const dy = point.y - cy;
-  const distance = Math.hypot(dx / rx, dy / ry);
-  if (distance <= 1) return point;
-  return {
-    x: cx + dx / distance,
-    y: cy + dy / distance
-  };
-}
-
-let sigmaCloudFilterSequence = 0;
-
-function nextSigmaCloudFilterSequence(): number {
-  sigmaCloudFilterSequence += 1;
-  return sigmaCloudFilterSequence;
-}
-
-function sigmaSharedCloudFilterDef(ownerDocument: Document, filterId: string): SVGSVGElement {
-  const svg = ownerDocument.createElementNS(SIGMA_OVERLAY_SVG_NS, "svg");
-  svg.setAttribute("aria-hidden", "true");
-  svg.style.position = "absolute";
-  svg.style.width = "0";
-  svg.style.height = "0";
-  svg.style.overflow = "hidden";
-  const defs = ownerDocument.createElementNS(SIGMA_OVERLAY_SVG_NS, "defs");
-  const filter = ownerDocument.createElementNS(SIGMA_OVERLAY_SVG_NS, "filter");
-  filter.setAttribute("id", filterId);
-  filter.setAttribute("x", "-50%");
-  filter.setAttribute("y", "-50%");
-  filter.setAttribute("width", "200%");
-  filter.setAttribute("height", "200%");
-  const blur = ownerDocument.createElementNS(SIGMA_OVERLAY_SVG_NS, "feGaussianBlur");
-  blur.setAttribute("stdDeviation", "20");
-  filter.append(blur);
-  defs.append(filter);
-  svg.append(defs);
-  return svg;
-}
-
-function sigmaCloudSvg(
-  ownerDocument: Document,
-  color: string,
-  cloud: SigmaCommunityCloud,
-  dim: boolean,
-  filterId: string,
-  onSelect: () => void
-): SVGSVGElement {
-  const svg = ownerDocument.createElementNS(SIGMA_OVERLAY_SVG_NS, "svg");
-  svg.setAttribute("width", "100%");
-  svg.setAttribute("height", "100%");
-  svg.style.position = "absolute";
-  svg.style.inset = "0";
-  svg.style.overflow = "visible";
-  svg.style.pointerEvents = "none";
-  let shape: SVGElement;
-  if (cloud.localPoints) {
-    const polygon = ownerDocument.createElementNS(SIGMA_OVERLAY_SVG_NS, "polygon");
-    polygon.setAttribute("points", cloud.localPoints.map((p) => `${p.x},${p.y}`).join(" "));
-    shape = polygon;
-  } else {
-    const ellipse = ownerDocument.createElementNS(SIGMA_OVERLAY_SVG_NS, "ellipse");
-    ellipse.setAttribute("cx", String(cloud.box.width / 2));
-    ellipse.setAttribute("cy", String(cloud.box.height / 2));
-    ellipse.setAttribute("rx", String(Math.max(8, cloud.box.width / 2)));
-    ellipse.setAttribute("ry", String(Math.max(8, cloud.box.height / 2)));
-    shape = ellipse;
-  }
-  shape.setAttribute("fill", color);
-  shape.setAttribute("fill-opacity", dim ? "0.06" : "0.2");
-  shape.setAttribute("filter", `url(#${filterId})`);
-  shape.style.pointerEvents = "fill";
-  shape.style.cursor = "pointer";
-  shape.addEventListener("click", (event) => {
-    event.stopPropagation();
-    onSelect();
-  });
-  svg.append(shape);
-  return svg;
-}
-
-function convexHull2d(points: GraphScreenPoint[]): GraphScreenPoint[] {
-  if (points.length < 3) return points.slice();
-  const pts = points.slice().sort((a, b) => a.x - b.x || a.y - b.y);
-  const cross = (o: GraphScreenPoint, a: GraphScreenPoint, b: GraphScreenPoint): number =>
-    (a.x - o.x) * (b.y - o.y) - (a.y - o.y) * (b.x - o.x);
-  const lower: GraphScreenPoint[] = [];
-  for (const p of pts) {
-    while (lower.length >= 2 && cross(lower[lower.length - 2], lower[lower.length - 1], p) <= 0) lower.pop();
-    lower.push(p);
-  }
-  const upper: GraphScreenPoint[] = [];
-  for (let i = pts.length - 1; i >= 0; i -= 1) {
-    const p = pts[i];
-    while (upper.length >= 2 && cross(upper[upper.length - 2], upper[upper.length - 1], p) <= 0) upper.pop();
-    upper.push(p);
-  }
-  lower.pop();
-  upper.pop();
-  return lower.concat(upper);
-}
-
-function overlayPointerScreenPoint(event: MouseEvent | PointerEvent, root: HTMLElement): GraphScreenPoint {
-  return rootClientPointToScreenPoint({
-    x: event.clientX,
-    y: event.clientY
-  }, root.getBoundingClientRect());
-}
-
-function sigmaScreenPointToWorldPoint(
-  sigma: SigmaGlobalSigmaLike,
-  point: GraphScreenPoint,
-  options: Pick<SigmaGlobalRendererCreateOptions, "viewport" | "viewportSize" | "adapterData">
-): { x: number; y: number } {
-  const projected = sigma.viewportToGraph?.(point);
-  if (projected && Number.isFinite(projected.x) && Number.isFinite(projected.y)) {
-    return projected;
-  }
-  return screenPointToWorldPoint(
-    point,
-    options.viewport ?? DEFAULT_RENDERER_VIEWPORT,
-    options.viewportSize ?? { width: 1, height: 1 },
-    options.adapterData.renderable.worldBounds
-  );
-}
-
-function sigmaWorldPointToScreenPoint(
-  sigma: SigmaGlobalSigmaLike,
-  point: { x: number; y: number },
-  options: Pick<SigmaGlobalRendererCreateOptions, "viewport" | "viewportSize" | "adapterData">
-): GraphScreenPoint {
-  const projected = sigma.graphToViewport?.(point);
-  if (projected && Number.isFinite(projected.x) && Number.isFinite(projected.y)) {
-    return projected;
-  }
-  const percent = worldPointToCssPercentPoint(point, options.adapterData.renderable.worldBounds);
-  const size = options.viewportSize ?? { width: 1, height: 1 };
-  return {
-    x: (percent.x / 100) * size.width,
-    y: (percent.y / 100) * size.height
-  };
 }
 
 function sigmaSettingsForTheme(theme: ThemeId): Record<string, unknown> {
